requirements.txt
black==25.9.0
click==8.3.0
mypy_extensions==1.1.0
numpy==2.3.4
packaging==25.0
pandas==2.3.3
pathspec==0.12.1
platformdirs==4.5.0
psycopg2-binary==2.9.11
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytokens==0.2.0
pytz==2025.2
six==1.17.0
tqdm==4.67.1
tzdata==2025.2
